I started playing with dwm again
conkyrc format changed to lua. There's a conversion script on the conky
project page but it requires some manual intervention so I opted to just
remake it. The colors don't seem to work still.

Add a scripts directory that should mostly hold things that the other
config files will need to run. This is likely only going to be dwm/i3
startup scripts. It has one script in it that powers the systray in a
default dwm install.